fix: resolve all clippy warnings failing CI (-D warnings)
- normalize.rs: use checked_div instead of manual zero-check division
- reading_order.rs: sort_by -> sort_by_key
- markdown_pipeline.rs: remove redundant reference in format! arg
- deskew, shadow, shutdown, web_integration: apply clippy suggestions

+++ superbook-pdf/src/margin/shadow.rs
@@ -718,7 +718,7 @@ mod tests {
 
         // Add shadow gradient on left edge
         for x in 0..50 {
-            let brightness = (128 + (x as u32 * 2)) as u8;
+            let brightness = (128 + (x * 2)) as u8;
             for y in 0..100 {
                 image.put_pixel(x, y, Rgb([brightness, brightness, brightness]));
             }
